+++ analyzers/PhishingInitiative/README.md
@@ -1,0 +1,9 @@
+### Phishing-Initiative 
+[Phishing-Initiative](https://phishing-initiative.fr/)  ables any Internet user to help fight against phishing attacks. When reporting us the address of a suspected phishing website, we’ll analyze it and have it blocked in the participating Web browsers.
+
+The analyzer comes in two flavours: lookup and scan. The first search in the database and can be used with basic API access while the second one requires higher profile role.
+
+#### Requirements
+You need a valid Phishing-Initiative API integration subscription to use the analyzer.
+
+- Provide your API key as values for the `key` parameter.
